move bitcoinUtils to bitcoin directory
