+++ src/ShareLogger.cc
@@ -28,6 +28,8 @@
 #include "Utils.h"
 #include "utilities_js.hpp"
 
+#include "bitcoin/BitcoinUtils.h"
+
 #include <algorithm>
 #include <string>
 
@@ -42,7 +44,6 @@
 #include <unistd.h>
 
 #include <chainparams.h>
-#include "BitcoinUtils.h"
 
 
 //////////////////////////////  ShareLogWriterT  ///////////////////////////////
