+++ test/TestStatistics.cc
@@ -25,8 +25,8 @@
 #include "gtest/gtest.h"
 #include "Common.h"
 #include "Statistics.h"
-#include "BitcoinUtils.h"
 
+#include "bitcoin/BitcoinUtils.h"
 #include "bitcoin/StratumBitcoin.h"
 
 ////////////////////////////////  StatsWindow  /////////////////////////////////
